Rename from Service to Client in package openai
That make sense as previously it's both client of that API and service of this app. But now since a new web Service of this app has been introduced, it shall be renamed to help distinguish which side the adaptor towards. Service towards human user while Client towards the opposite.