rename id to sessionID in args of chat.Service methods
Previously, user have to guess it from the fact that there is only session id and chat id is not generated yet at invoked stage. But now the name Chat Service is error-prone to leads to the idea that id is chat's. So I change it to make it clear.