omit further ID-like fields in JSON
Similar to commit e9d78647.
I found that in fulfilled Session, Chat and Result are leaking information that I want to hide between users.
At present none of those object are editable through APIs, so just omit now, rather to accomplish an elegant way that distinguish the admin and user view, likes that in Session's ID.